requirements.txt
fastapi
uvicorn
groq
python-dotenv
pydantic
python-multipart
motor
dnspython
shutil
marked
requests